mnist_app/requirements.txt
streamlit
numpy
pillow
scikit-learn
opencv-python-headless
